fix: rename typescript files to use camelCase.ts
